server/requirements.txt
django
djangorestframework
coreapi
django-rest-swagger
django-cors-headers